backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# YouTube
yt-dlp>=2024.12.23
pytube==15.0.0

# AI
google-generativeai>=0.8.0
openai==1.12.0

# Audio Processing
openai-whisper==20231117
pydub==0.25.1

# Video Processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# HTTP & APIs
httpx==0.27.0
requests==2.31.0
spotipy==2.23.0

# Database
supabase==2.3.4

# Async
celery==5.3.6
redis==5.0.1

# Utilities
python-multipart==0.0.6
